# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Snaglist implemented platform API",
    "version": "0.3.0-candidate",
    "description": "Current implemented v2 foundation, not the complete unified-platform API. Native full graph sync, canonical evidence/workflow/media, reports/jobs, company billing and account deletion remain under implementation. V1 recovery endpoints remain outside this contract. Dates/money in this first snag slice retain legacy timestamp/number DTOs pending explicit date-only/decimal migration; do not claim completed field parity."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment; explicit PORTAL_ORIGIN and PLATFORM_ENVIRONMENT on server."
    }
  ],
  "paths": {
...
    "/api/v2/projects/{projectId}/snags/{snagId}": {
      "get": {
        "operationId": "getSnag",
        "summary": "Read a snag including explicit archive state",
        "description": "Authenticated user",
        "security": [
          {
            "browserSession": []
          },
          {
            "nativeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "snagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Committed result. Mutations return only after transaction commit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSnagResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign in again",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Action not permitted or previously held project access revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable without revealing cross-workspace existence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Revision, identity, operation or lifecycle conflict. Retain the original draft/request.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Error"
                    },
                    {
                      "$ref": "#/components/schemas/SnagRevisionConflict"
                    }
                  ]
                }
              }
            }
          },
          "410": {
            "description": "Expired, archived or rebootstrap required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Bounded request limit reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Temporary provider/service failure; do not infer success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "editSnag",
        "summary": "Edit ordinary fields with a base revision",
        "description": "Project edit; deadline changes require assign permission. Absent means unchanged, explicit null clears. Closed state/assignment use separate actions; no last-writer-wins. Calendar deadlines (dueOn or deprecated dueDate) require Manager assignment rights.",
        "security": [
          {
            "browserSession": []
          },
          {
            "nativeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "snagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exact PORTAL_ORIGIN; required for cookie-authenticated mutations and public browser email exchange."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Required with browser-session cookies. Native Bearer requests do not use CSRF."
          }
        ],
        "responses": {
          "200": {
            "description": "Committed result. Mutations return only after transaction commit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSnagResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign in again",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Action not permitted or previously held project access revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable without revealing cross-workspace existence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Revision, identity, operation or lifecycle conflict. Retain the original draft/request.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Error"
                    },
                    {
                      "$ref": "#/components/schemas/SnagRevisionConflict"
                    }
                  ]
                }
              }
            }
          },
          "410": {
            "description": "Expired, archived or rebootstrap required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Bounded request limit reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Temporary provider/service failure; do not infer success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnagEditCommand"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "browserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-snaglist_session",
        "description": "Secure, HttpOnly, SameSite=Lax, host-only. Unsafe calls require exact Origin and X-CSRF-Token."
      },
      "nativeBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Signature, expiry, subject/user UUID, active lifecycle and revocation version checked."
      }
    },
    "schemas": {
...
      "MutationMetadata": {
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "deviceId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "operationId",
          "deviceId"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "reason": {
            "type": "string"
          },
          "identifier": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "reason",
          "identifier"
        ]
      },
...
      "Snag": {
        "type": "object",
        "properties": {
          "reference": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "open",
              "in_progress",
              "awaiting_review",
              "changes_requested",
              "closed"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "closedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "assignedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "contractorId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "tradeId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "drawingId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "costEstimate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ]
          },
          "actualCost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ]
          },
          "drawingPinX": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ]
          },
          "drawingPinY": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "projectId",
          "ownerId",
          "reference",
          "title",
          "status",
          "priority",
          "currency",
          "tags"
        ]
      },
      "PlatformSnagResponse": {
        "type": "object",
        "properties": {
          "snag": {
            "$ref": "#/components/schemas/Snag"
          },
          "revision": {
            "type": "integer",
            "minimum": 1
          },
          "workflowRevision": {
            "type": "integer",
            "minimum": 1
          },
          "displayNumber": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ]
          },
          "publishedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "archivedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "archiveReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "canonical": {
            "$ref": "#/components/schemas/CanonicalSnagValues",
            "description": "Authoritative exact dates/costs. The nested snag retains compatibility timestamp/number mirrors. Missing only in stored pre-parity candidate receipts or snapshots: refetch before presenting/editing these fields."
          }
        },
        "required": [
          "snag",
          "revision",
          "workflowRevision"
        ]
      },
...
      "SnagRevisionConflict": {
        "type": "object",
        "properties": {
          "error": {
            "const": true
          },
          "identifier": {
            "const": "revision_conflict"
          },
          "reason": {
            "type": "string"
          },
          "current": {
            "$ref": "#/components/schemas/PlatformSnagResponse"
          },
          "changedFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "error",
          "identifier",
          "reason",
          "current",
          "changedFields"
        ]
      },
      "SnagFields": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ]
          },
          "currency": {
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "deprecated": true,
            "description": "Earlier unreleased candidate compatibility only. Use dueOn or the corresponding Decimal string field. Supplying both representations is rejected; no excess precision is rounded."
          },
          "costEstimate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "deprecated": true,
            "description": "Earlier unreleased candidate compatibility only. Use dueOn or the corresponding Decimal string field. Supplying both representations is rejected; no excess precision is rounded."
          },
          "actualCost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "deprecated": true,
            "description": "Earlier unreleased candidate compatibility only. Use dueOn or the corresponding Decimal string field. Supplying both representations is rejected; no excess precision is rounded."
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dueOn": {
            "anyOf": [
              {
                "type": "string",
                "format": "date",
                "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
              },
              {
                "type": "null"
              }
            ]
          },
          "costEstimateDecimal": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(0|[1-9][0-9]{0,9})(\\.[0-9]{1,6})?$",
                "description": "Exact decimal major currency units, 0\u20131000000000, up to six decimal places. Do not parse into a JavaScript float for edits. Null means unset; \"0\" means a recorded zero."
              },
              {
                "type": "null"
              }
            ]
          },
          "actualCostDecimal": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(0|[1-9][0-9]{0,9})(\\.[0-9]{1,6})?$",
                "description": "Exact decimal major currency units, 0\u20131000000000, up to six decimal places. Do not parse into a JavaScript float for edits. Null means unset; \"0\" means a recorded zero."
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [],
        "additionalProperties": false,
        "minProperties": 1,
        "description": "Only ordinary data fields; deadline fields require Manager rights. Absent is unchanged, null clears. Due dates stay calendar-only in the workspace timezone. Assignment/workflow are separate commands."
      },
...
      "SnagEditCommand": {
        "type": "object",
        "properties": {
          "mutation": {
            "$ref": "#/components/schemas/MutationMetadata"
          },
          "expectedRevision": {
            "type": "integer",
            "minimum": 1
          },
          "fields": {
            "$ref": "#/components/schemas/SnagFields"
          }
        },
        "required": [
          "mutation",
          "expectedRevision",
          "fields"
        ]
      },
...
      "CanonicalSnagValues": {
        "type": "object",
        "properties": {
          "dueOn": {
            "anyOf": [
              {
                "type": "string",
                "format": "date",
                "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
              },
              {
                "type": "null"
              }
            ]
          },
          "costEstimateDecimal": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(0|[1-9][0-9]{0,9})(\\.[0-9]{1,6})?$",
                "description": "Exact decimal major currency units, 0\u20131000000000, up to six decimal places. Do not parse into a JavaScript float for edits. Null means unset; \"0\" means a recorded zero."
              },
              {
                "type": "null"
              }
            ]
          },
          "actualCostDecimal": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(0|[1-9][0-9]{0,9})(\\.[0-9]{1,6})?$",
                "description": "Exact decimal major currency units, 0\u20131000000000, up to six decimal places. Do not parse into a JavaScript float for edits. Null means unset; \"0\" means a recorded zero."
              },
              {
                "type": "null"
              }
            ]
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          }
        },
        "required": [
          "currency"
        ]
      }
    }
  }
}